+++ quotas/jinja2/quota-definitions/bulk-create-definition-intro-period.jinja
@@ -1,0 +1,4 @@
+{# 
+Displays if the option to include an intro period has been selected
+Duplicate of the current create definition period page.
+#}
