+++ quotas/jinja2/quota-definitions/bulk-create-definitions-create.jinja
@@ -1,0 +1,6 @@
+{# 
+This will be an update of the current definition create template
+Should also include 
+a numerical input for the number of deffinition periods to be created
+an option collecing the date split for the different definition periods 
+ #}
