+++ quotas/jinja2/quota-definitions/bulk-create-definitions-start.jinja
@@ -1,0 +1,4 @@
+{# 
+First step for bulk creating definition periods
+Collects whether there's an introductory period
+#}
